pyproject.toml
[project]
name = "ninja-robot-v2"
version = "0.1.0"
description = "NinjaRobot V2 - Raspberry Pi Zero 2W Robot"
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "google-generativeai>=0.3.0",
    "SpeechRecognition>=3.10.0",
    "gTTS>=2.5.0",
    "pygame>=2.5.0",
    "google-cloud-speech>=2.20.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "RPi.GPIO>=0.7.1 ; sys_platform == 'linux'",
    "smbus2>=0.4.3 ; sys_platform == 'linux'",
]

[dependency-groups]
dev = [
    "pylint>=3.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
    "pytest>=8.0.0",
    "types-requests",
    "types-setuptools",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ninja_robot"]


[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pylint.messages_control]
disable = [
    "C0111", # missing-docstring (we'll add them gradually)
    "C0103", # invalid-name (for now, to avoid noise with existing names if any)
]

[tool.black]
line-length = 88
target-version = ['py311']
